refactor: move data into modules/character
