+++ src/modules/character/data.ts
